% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{bin}
\alias{bin}
\title{Bin numeric vector}
\usage{
bin(x, d, bmin = 0, bmax = max(x, na.rm = T), ...)
}
\arguments{
\item{x}{numeric vector}

\item{d}{bin width}

\item{bmin}{minimum bin (defaults to \code{0})}

\item{bmax}{maximum bin (defaults to \code{max(x,na.rm=T)})}

\item{...}{additional arguments passed to cut}
}
\value{
factor, indicating bin assignment
}
\description{
Bin numeric vector
}
