% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{convert_digital_size}
\alias{convert_digital_size}
\title{Convert digital size}
\usage{
convert_digital_size(ds)
}
\arguments{
\item{ds}{numeric}
}
\value{
numeric
}
\description{
Apply the calibration function from the Focal OPC manual (Appendix A)
to convert from raw 1-4096 digital size bins recorded by the OPC to
Equivalent Spherical Diameter (ESD) in mm. This is called within
\code{convert_single_opc()}. It was modified from Mark Baumgartner's IDL
routine of the same name.
}
