% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{convert_single_opc}
\alias{convert_single_opc}
\title{Convert OPC data}
\usage{
convert_single_opc(ifile)
}
\arguments{
\item{ifile}{path to .D00 file}
}
\value{
tibble
}
\description{
Read raw OPC data and convert to nice tabular format. This was modified (heavily) from
from Mark Baumgartner's IDL routine of the same name.
}
