% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{drop_volume}
\alias{drop_volume}
\title{Drop volume}
\usage{
drop_volume(df, dz, area = 0.02 * 0.25)
}
\arguments{
\item{df}{OPC tibble}

\item{dz}{depth bin width}

\item{area}{area (m2) of OPC tunnel (defaults to \code{0.02*0.25})}
}
\value{
df (OPC tibble)
}
\description{
Identify depth bins where less than 50\% of the bin volume was sampled, and set corresponding
particle concentration to \code{NA}. This avoids large spikes in concentration caused by poorly sampled bins.
}
