% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\docType{data}
\name{opc}
\alias{opc}
\title{Processed OPC downcast}
\format{
A nested tibble with 148 rows and 8 variables:
\describe{
\item{scan}{the scan number, which iteratively increases with each data record}
\item{timer}{the timer, which recorded seconds since the unit was powered on}
\item{atten}{light attenuation}
\item{depth}{instrument depth in meters}
\item{flag}{quality control flag, with zero meaning good. See \code{opc_flag()} for the other definitions}
\item{time}{the datetime since deployment, in UTC}
\item{volume_filtered}{the volume of water that has passed through the OPC since the previous record, in cubic meters}
\item{esd}{a nested list of the particle sizes, in Equivalent Spherical Diameter (ESD; mm) detected during this data record}
}
}
\source{
hansen.johnson@dal.ca
}
\usage{
opc
}
\description{
A dataset containing a processed, trimmed OPC downcast (cast id \verb{2019b_27}) that
was collected in the southern Gulf of St Lawrence in August, 2019.
}
\keyword{datasets}
