% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_biomass}
\alias{opc_biomass}
\title{Calculate OPC biomass}
\usage{
opc_biomass(df, dz = 2, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (meters)}

\item{good_only}{use only good (unflagged) values}
}
\value{
tibble
}
\description{
Calculate OPC biomass
}
