% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_check}
\alias{opc_check}
\title{Check opc data}
\usage{
opc_check(df)
}
\arguments{
\item{df}{OPC tibble}
}
\value{
df (OPC tibble)
}
\description{
Simple subroutine to catch common plotting / processing errors
}
