% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_energy}
\alias{opc_energy}
\title{Calculate OPC energy}
\usage{
opc_energy(df, dz = 2, good_only = TRUE, reject_volume = TRUE)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (meters)}

\item{good_only}{use only good (unflagged) values}

\item{reject_volume}{set concentration to \code{NA} for depth bins with <50\% volume sampled}
}
\value{
tibble
}
\description{
Convert Equivalent Spherical Diameter (ESD) from the OPC into
energy (in Joules) per depth bin using the formula below which was
developed by Davies et al (2012).
}
\details{
\code{energy = 0.0134 * (esd) - 7.52}
}
