% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_flag}
\alias{opc_flag}
\title{Flag bad data}
\usage{
opc_flag(df)
}
\arguments{
\item{df}{OPC data frame}
}
\value{
OPC tibble with updated flag column
}
\description{
Flag bad OPC data based on several criteria. When a record is flagged
the \code{flag} value is updated to reflect the reason. The criteria and
associated flag values are as follows:
}
\details{
\enumerate{
\item extreme depths - \code{depth}
\item non-sequential timer values - \code{timer}
\item slow descent rates (<0.3 m/s) - \code{slow}
\item directional reversals - \code{reversal}
\item extreme light attenuance - \code{attenuance}
\item extreme volume filtered - \code{volume}
}
}
