% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_histogram}
\alias{opc_histogram}
\title{Calculate OPC size-frequency histogram}
\usage{
opc_histogram(df, ds = 0.05, min_size = 1, max_size = 4, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{ds}{particle size bin width (mm)}

\item{min_size}{minimum particle ESD (mm)}

\item{max_size}{maximum particle ESD (mm)}

\item{good_only}{use only good (unflagged) values}
}
\value{
tibble
}
\description{
Calculate OPC size-frequency histogram
}
