% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_abundance}
\alias{opc_plot_abundance}
\title{Compute and plot OPC abundance vs depth}
\usage{
opc_plot_abundance(df, dz = 4, min_size = 1, max_size = 4, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (m)}

\item{min_size}{minimum particle ESD (mm)}

\item{max_size}{maximum particle ESD (mm)}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Compute and plot OPC abundance vs depth
}
