% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_attenuance}
\alias{opc_plot_attenuance}
\title{Plot OPC attenuance versus depth}
\usage{
opc_plot_attenuance(df, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Plot OPC attenuance versus depth
}
