% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_biomass}
\alias{opc_plot_biomass}
\title{Compute and plot OPC biomass vs depth}
\usage{
opc_plot_biomass(df, dz = 4, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (m)}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Compute and plot OPC biomass vs depth
}
