% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_depth}
\alias{opc_plot_depth}
\title{Plot OPC scan versus depth series}
\usage{
opc_plot_depth(df, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Plot OPC scan versus depth series
}
