% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_diagnostics}
\alias{opc_plot_diagnostics}
\title{Plot a suite of OPC diagnostics}
\usage{
opc_plot_diagnostics(
  df,
  dz = 2,
  ds = 0.05,
  min_size = 0,
  max_size = 4,
  amin_size = 1,
  amax_size = 2,
  good_only = T
)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (m)}

\item{ds}{particle size bin width (mm)}

\item{min_size}{minimum particle ESD (mm) for full plot}

\item{max_size}{maximum particle ESD (mm) for full plot}

\item{amin_size}{minimum particle ESD (mm) for abundance profile}

\item{amax_size}{maximum particle ESD (mm) for abundance profile}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Plot a suite of OPC diagnostics
}
