% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_energy}
\alias{opc_plot_energy}
\title{Compute and plot OPC energy vs depth}
\usage{
opc_plot_energy(df, dz = 4, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (m)}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Compute and plot OPC energy vs depth
}
