% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_image}
\alias{opc_plot_image}
\title{Compute and plot image-style plot of OPC abundance in size and depth bins}
\usage{
opc_plot_image(
  df,
  ds = 0.05,
  dz = 2,
  min_size = 0,
  max_size = 4,
  good_only = T
)
}
\arguments{
\item{df}{opc tibble}

\item{ds}{particle size bin width (mm)}

\item{dz}{depth bin width (m)}

\item{min_size}{minimum particle ESD (mm)}

\item{max_size}{maximum particle ESD (mm)}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Compute and plot image-style plot of OPC abundance in size and depth bins
}
