% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_plot_time_count}
\alias{opc_plot_time_count}
\title{Plot timeseries of OPC particle counts in given size range}
\usage{
opc_plot_time_count(df, dt = 1, min_size = 1, max_size = 4, good_only = TRUE)
}
\arguments{
\item{df}{opc tibble}

\item{dt}{timestep (s)}

\item{min_size}{minimum particle ESD (mm)}

\item{max_size}{maximum particle ESD (mm)}

\item{good_only}{use only good (unflagged) values}
}
\value{
ggplot
}
\description{
Plot timeseries of OPC particle counts in given size range
}
