% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_process_benchtest}
\alias{opc_process_benchtest}
\title{Benchtest a single OPC cast}
\usage{
opc_process_benchtest(ifile)
}
\arguments{
\item{ifile}{path to .D00 file}
}
\value{
opc tibble
}
\description{
The processing occurs in several steps:
\enumerate{
\item read in binary OPC data in .D00 file with \code{read_focal_opc()}
\item convert to a nice tabular format with \code{convert_single_opc()}
\item use a shiny app to interactively check the results with \code{opc_trim_benchtest()}
}
}
