% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_process_cruise}
\alias{opc_process_cruise}
\title{Process multiple OPC casts}
\usage{
opc_process_cruise(cruise, data_dir, output_dir = data_dir, overwrite = FALSE)
}
\arguments{
\item{cruise}{prefix to add to each processed data file}

\item{data_dir}{path to raw (D00) data files}

\item{output_dir}{path to processed data files}

\item{overwrite}{overwrite processed data? (default is FALSE)}
}
\value{
OPC tibble
}
\description{
Process all the OPC raw data files in \code{data_dir} and save processed downcasts in
the \code{output_dir}. The naming convention extracts the cast number from the D00 file
name, and saves each downcast according to the convention:
(output_dir)/(cruise)_(cast).rds
}
