% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_process_multicast}
\alias{opc_process_multicast}
\title{Process multiple OPC casts in single file}
\usage{
opc_process_multicast(ifile, cast_ids)
}
\arguments{
\item{ifile}{path to raw data file (*.D00)}

\item{cast_ids}{vector indicating the IDs used for each extracted cast}
}
\value{
OPC tibble
}
\description{
Occasionally multiple OPC casts are conducted repeatedly in the same
data file. This function runs \code{opc_process()} multiple times on the same file
to allow the user to select multiple casts.
}
