% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_speed}
\alias{opc_speed}
\title{Calculate OPC speed}
\usage{
opc_speed(df, method = "instantaneous", exclude_bad_depths = TRUE)
}
\arguments{
\item{df}{opc tibble}

\item{method}{choose one from \code{instantaneous}, \code{range}, or \code{lm}}

\item{exclude_bad_depths}{logical, where \code{TRUE} removes bad depth flags}
}
\value{
numeric speed in meters per second (positive down)
}
\description{
Use one of 3 methods (\code{instantaneous}, \code{range}, or \code{lm}) to calculate the
descent rate of an OPC in meters per second.
}
\details{
\code{instantaneous} - compute speed at each sampling point and return average
\code{range} - simple divide the time difference by the depth difference
\code{lm} - fit a simple linear model and return the slope parameter
}
