% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_trim}
\alias{opc_trim}
\title{Trim OPC cast}
\usage{
opc_trim(df)
}
\arguments{
\item{df}{OPC tibble}
}
\value{
OPC tibble trimmed to include selected downcast
}
\description{
Shiny app to select the downcast portion of an OPC cast. It also applies
\code{opc_flag()} to flag bad values and uses \code{opc_plot_diagnostics()} to produce
a helpful diagnostic plot of the selected region.
}
