% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_trim_bechtest}
\alias{opc_trim_bechtest}
\title{Benchtest OPC cast}
\usage{
opc_trim_bechtest(df)
}
\arguments{
\item{df}{OPC tibble}
}
\value{
OPC tibble trimmed to include selected time
}
\description{
Shiny app to display timeseries and histogram of OPC cast
}
