% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{read_focal_opc}
\alias{read_focal_opc}
\title{Read OPC raw data file}
\usage{
read_focal_opc(
  ifile,
  nheader = 27,
  tformat = "WHOI \%m/\%d/\%y \%H:\%M:\%S \\n\\n\\r",
  tz = "UTC"
)
}
\arguments{
\item{ifile}{path to .D00 file}

\item{nheader}{number of bits in header (default = 27)}

\item{tformat}{strptime format used to extract timestamp from header}

\item{tz}{timezone}
}
\value{
list containing \code{start_time} of the data file and vectors of the \code{id} flags and
corresponding \code{data} values
}
\description{
OPC data are stored in a unique binary file format with *.D00 file
extension. This function was adapted from Mark Baumgartner's IDL routine
of the same name to read data from a *.D00 file and extract the id flags
and data fields.
}
