% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{relabel}
\alias{relabel}
\title{Relabel numeric bin}
\usage{
relabel(x)
}
\arguments{
\item{x}{factor indicating bin assignment}
}
\value{
numeric, indicating left hand side of bin interval (\code{a} of \verb{(a,b]})
}
\description{
Relabel numeric bin
}
