% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_abundance}
\alias{opc_abundance}
\title{Calculate OPC abundance profile}
\usage{
opc_abundance(df, dz = 2, min_size = 1, max_size = 4, good_only = TRUE)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (meters)}

\item{min_size}{minimum particle ESD (mm)}

\item{max_size}{maximum particle ESD (mm)}

\item{good_only}{use only good (unflagged) values}
}
\value{
tibble
}
\description{
Calculate OPC abundance profile
}
\examples{

data(opc)

}
