% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_biomass}
\alias{opc_biomass}
\title{Calculate OPC biomass}
\usage{
opc_biomass(df, dz = 2, good_only = T)
}
\arguments{
\item{df}{opc tibble}

\item{dz}{depth bin width (meters)}

\item{good_only}{use only good (unflagged) values}
}
\value{
tibble
}
\description{
Convert Equivalent Spherical Diameter (ESD) from the OPC into
wet weight (in mg) per depth bin using the formula below which was
developed by Suthers et al (2006) and implemented by
Fortune et al (2020).
}
\details{
\code{wet_weight = 4/3 * pi * (esd/2)^3 * rho}
where \code{rho} = 1 mg / mm

Note that this is only applied to particles > 1 mm ESD
}
