% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opc.R
\name{opc_process}
\alias{opc_process}
\title{Process a single OPC cast}
\usage{
opc_process(ifile)
}
\arguments{
\item{ifile}{path to .D00 file}
}
\value{
opc tibble
}
\description{
The processing occurs in several steps:
\enumerate{
\item read in binary OPC data in .D00 file with \code{read_focal_opc()}
\item convert to a nice tabular format with \code{convert_single_opc()}
\item apply various quality control flags with \code{opc_flag()}
\item use a shiny app to interactively select the downcast with \code{opc_trim()}
}
}
